Moved DeepStruct to separate file
